redis / jpa separate repositories
