+++ src/main/java/kr/pincoin/durian/auth/service/RoleService.java
@@ -1,7 +1,7 @@
 package kr.pincoin.durian.auth.service;
 
 import kr.pincoin.durian.auth.domain.Role;
-import kr.pincoin.durian.auth.repository.RoleRepository;
+import kr.pincoin.durian.auth.repository.jpa.RoleRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
