--- src/main/java/kr/pincoin/durian/shop/repository/CategoryRepositoryQuery.java
@@ -1,4 +1,0 @@
-package kr.pincoin.durian.shop.repository;
-
-public interface CategoryRepositoryQuery {
-}
