--- src/main/java/kr/pincoin/durian/shop/repository/ProductRepositoryQuery.java
@@ -1,4 +1,0 @@
-package kr.pincoin.durian.shop.repository;
-
-public interface ProductRepositoryQuery {
-}
