--- src/main/java/kr/pincoin/durian/shop/repository/VoucherRepositoryQuery.java
@@ -1,4 +1,0 @@
-package kr.pincoin.durian.shop.repository;
-
-public interface VoucherRepositoryQuery {
-}
