+++ src/main/java/kr/pincoin/durian/shop/repository/jpa/CategoryRepository.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.shop.repository;
+package kr.pincoin.durian.shop.repository.jpa;
 
 import kr.pincoin.durian.shop.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
