+++ src/main/java/kr/pincoin/durian/shop/repository/jpa/CategoryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.shop.repository;
+package kr.pincoin.durian.shop.repository.jpa;
 
 public class CategoryRepositoryImpl implements  CategoryRepositoryQuery {
 }
