+++ src/main/java/kr/pincoin/durian/shop/repository/jpa/CategoryRepositoryQuery.java
@@ -1,0 +1,4 @@
+package kr.pincoin.durian.shop.repository.jpa;
+
+public interface CategoryRepositoryQuery {
+}
