+++ src/main/java/kr/pincoin/durian/shop/repository/jpa/CategoryTreePathRepositoryImpl.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.shop.repository;
+package kr.pincoin.durian.shop.repository.jpa;
 
 public class CategoryTreePathRepositoryImpl implements  CategoryRepositoryQuery {
 }
