+++ src/main/java/kr/pincoin/durian/shop/repository/jpa/ProductRepositoryImpl.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.shop.repository;
+package kr.pincoin.durian.shop.repository.jpa;
 
 public class ProductRepositoryImpl implements  ProductRepositoryQuery {
 }
