+++ src/main/java/kr/pincoin/durian/shop/repository/jpa/VoucherRepository.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.shop.repository;
+package kr.pincoin.durian.shop.repository.jpa;
 
 import kr.pincoin.durian.shop.domain.Voucher;
 import org.springframework.data.jpa.repository.JpaRepository;
